[HJ] fix : RestAPI와 통근버스 메인페이지 연결

+++ DigicoBus/app/src/main/java/com/kt/digicobus/dialog/BottomSheetStopLocation.kt
@@ -67,7 +67,7 @@ class BottomSheetStopLocation(val selectBusInfo:CommuteBusInfo) :
     override fun onMapReady(naverMap: NaverMap) {
         val naverMapAPIService = NaverMapAPIService(naverMap)
 
-        val latlng = LatLng(selectBusInfo.latitude!!, selectBusInfo.longitude!!)
+        val latlng = LatLng(selectBusInfo.latitude!!.toDouble(), selectBusInfo.longitude!!.toDouble())
         // camera position
         naverMapAPIService.setCameraPosition(latlng)
